refactor: moved fetch requests outside of components
